+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+#### 0.1.2
+* 04/04/2024
+* Changed the constructor to a class.
+* Added a max precision option. If a number can be represented exactly will a smaller exponent, it will do so, but it will not exceed precision.
+* Strings with decimal values are now allowed in the constructor parameters.
+* Merged the multiply methods into one.
+
 #### 0.1.1
 * 04/01/2024 `setExponent` now rounds. Added `getBitAtPositionFromRight`.